% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_mat_qn.R
\name{covMatQn}
\alias{covMatQn}
\title{Robust covariance matrix}
\usage{
covMatQn(x)
}
\arguments{
\item{x}{a numeric matrix}
}
\value{
a numeric matrix
}
\description{
Computes the robust covariance matrix of the matrix \code{x} proposed by Ma and Genton (2001) using the robust scale Qn of Rousseeuw and Croux (1993).
}
\examples{
dataset <- cbind(rnorm(100),rnorm(100))
covMatQn(dataset)
}
\references{
Ma, Y. and Genton, M. G. (2001) Highly robust estimation of dispersion matrices. \emph{Journal of Multivariate Analysis}, \bold{78}, 11--36.

Rousseeuw, P. J. and Croux, C. (1993) Alternatives to the median absolute deviation. \emph{Journal of the American Statistical Association}, \bold{88}, 1273--1283.
}
