% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_qn.R
\name{covQn}
\alias{covQn}
\title{Robust covariance between the variables \code{x} and \code{y}}
\usage{
covQn(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\value{
a numerical value with the robust covariance between \code{x} and \code{y}
}
\description{
Computes the robust covariance of \code{x} and \code{y} proposed by Ma and Genton (2001) using the robust scale Qn of Rousseeuw and Croux (1993).
}
\examples{
covQn(rnorm(100),rnorm(100))
}
\references{
Ma, Y. and Genton, M. G. (2001) Highly robust estimation of dispersion matrices. \emph{Journal of Multivariate Analysis}, \bold{78}, 11--36.

Rousseeuw, P. J. and Croux, C. (1993) Alternatives to the median absolute deviation. \emph{Journal of the American Statistical Association}, \bold{88}, 1273--1283.
}
